spring security and tomcat redirect 10/2

+++ src/main/java/com/ucrisko/libroomreserve/models/User.java
@@ -18,24 +18,24 @@ public class User {
   @Id
   @Column
   @GeneratedValue(strategy=GenerationType.AUTO)
-  private int userId;
+  private Long userId;
   
   @Column
   private String userName;
   
   public User(){}
   
-  public User(int userId, String userName){
+  public User(Long userId, String userName){
     super();
     this.userId = userId;
     this.userName = userName;
   }
 
-  public int getUserId() {
+  public Long getUserId() {
     return userId;
   }
 
-  public void setUserId(int userId) {
+  public void setUserId(Long userId) {
     this.userId = userId;
   }
